feat(http): move http methods into core module
